backend/requirements.txt
gunicorn
fastapi
uvicorn
google-cloud-firestore
google-auth
google-cloud-aiplatform
